Looking at the behaviour of GitHub Actions, what takes the most time is to wait for a runner for a workflow.
Once the workflow has a runner, we spend less time waiting for executors for each job.

As a consequence, it makes sense to limit the number of *workflows* as much as possible, and to use the *jobs* for a proper reporting in GitHub.

This commit is an attempt to have one workflow for PR, and one workflow for push.

It implies some duplication of code but I guess this is the price to pay for less wait time.