[3166] Incorporated examples into procedures
The contents of the following example files
modules/end-user-guide/examples/che-ref_a-kubernetes-secret-as-a-file.adoc
modules/end-user-guide/examples/che-ref_a-kubernetes-secret-as-an-variable.adoc

have been incorporated into the following procedure files
modules/end-user-guide/partials/proc_mounting-a-secret-as-a-file-into-a-workspace-container.adoc
modules/end-user-guide/partials/proc_mounting-a-secret-as-an-environment-variable-into-a-workspace-container.adoc

The two example files have been deleted.
The two procedure files have been rewritten.

Bug: https://issues.redhat.com/browse/RHDEVDOCS-3166