Set SHELL env var in devfile.yaml to allow terminal use
The 'quay.io/eclipse/che-docs:next' image does not use the usual
entrypoint (that adds an entry to /etc/passwd). As a result, the default
shell for the current user on OpenShift is /sbin/nologin, which prevents
opening terminal in the Code editor.

Since the image has Bash 5.2 installed, we can set the SHELL environment
variable, which will cause Code to use that shell for terminals.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>