Add caution against installing operators in separate namespaces
Add a warning to the installation docs to warn against installing the {prod}
Operator in a different namespace than the {devworkspace} Operator (operator
and any operators it depends on must be in the same namespace).

In addition, update the {devworkspace} variable to reflect accurate name for
DevWorkspace (no space between Dev Workspace).

Update modules/administration-guide/pages/installing-che-on-openshift-using-the-web-console.adoc

Co-authored-by: Jana Vrbkova <jvrbkova@redhat.com>
Signed-off-by: Angel Misevski <amisevsk@redhat.com>